create a separate .py file called streamlit_logs
